% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptRidge.R
\name{adaptRidge}
\alias{adaptRidge}
\title{adaptRidge}
\usage{
adaptRidge(
  Freq = Freq,
  phenotypes = phenotypes,
  clusters = clusters,
  eta = 1,
  max_iter = 1000,
  tol = 1e-08,
  seed = 1,
  nb.batch = 1,
  trace = FALSE
)
}
\arguments{
\item{Freq}{data frame of allele frequencies}

\item{phenotypes}{data frame of phenotypes}

\item{clusters}{pool ID}

\item{eta}{default=1}

\item{max_iter}{float default=1e3}

\item{tol}{float default=1e-8}

\item{seed}{default=1}

\item{nb.batch}{float default=1}

\item{trace}{boolean default=FALSE}
}
\value{
numeric vector of SNP effects
}
\description{
adaptative Ridge method with descending gradient
}
\examples{
sim = PhenoSim ( 1000, 100, 10, 0.6, 5 )
K = 7
clusters = quantile_clustering ( data.frame ( sim$phenotypes ), K )
Freq = compute_group_MAFs( sim$genotypes, as.factor(clusters) )
res = adaptRidge( Freq, sim$phenotypes[,1], clusters )
plot ( res$criterion )
beta = res$beta
pvalues<- pvalues.from.test.stat(beta)

}
