% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchPooGwaM.R
\name{benchPooGwaM}
\alias{benchPooGwaM}
\title{benchPooGwaM}
\usage{
benchPooGwaM(pvalues, causal.true, threshold = 0.1)
}
\arguments{
\item{pvalues}{data frame of phenotypes}

\item{causal.true}{numeric vector of the clusters}

\item{threshold}{float default=0.1}
}
\value{
list benchmarking metrics
}
\description{
benchmarking performance of PooGwaM on simulated data
}
\examples{
sim = PhenoSim (500, 100, 10, 0.6, 2)
causal.true = sim$causal.true
K = 4
clusters = quantile_clustering (data.frame(sim$phenotypes), K)
Freq= compute_group_MAFs(sim$genotypes,as.factor(clusters))
test.stat<-GWalpha(sim$phenotypes, d=1, clusters, Freq)
benchPooGwaM (test.stat, causal.true)


}
