% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chisquare.R
\name{Chisquare}
\alias{Chisquare}
\title{Chisquare}
\usage{
Chisquare(genotypes, clusters)
}
\arguments{
\item{genotypes}{data frame of genotypes}

\item{clusters}{numeric vector of the clusters}
}
\value{
list
}
\description{
apply classical chi-square
}
\examples{
sim = PhenoSim (1000, 100, 10, 0.6, 5)
clusters = quantile_clustering (sim$phenotypes, 7)
pv <- Chisquare (sim$genotypes, clusters)


}
