% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCausalSNP.R
\name{getCausalSNP}
\alias{getCausalSNP}
\title{Get causal SNPs from the simulation}
\usage{
getCausalSNP(simulation)
}
\arguments{
\item{simulation}{PhenotypeSimulator object}
}
\value{
list
}
\description{
Get causal SNPs from the simulation
}
\examples{
sim = sim2snp(100,100,10)
causalSNP = getCausalSNP(sim)
}
