% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAlpha.R
\name{GWalpha}
\alias{GWalpha}
\title{GWAlpha}
\usage{
GWalpha(
  phenotypes = phenotypes,
  clusters = clusters,
  Freq = Freq,
  K = K,
  penalized = TRUE
)
}
\arguments{
\item{phenotypes}{data frame of phenotypes}

\item{clusters}{numeric vector of the clusters}

\item{Freq}{numeric vector of allele frequencies}

\item{K}{numeric vector of number of clusters}

\item{penalized}{boolean}
}
\value{
numeric vector
}
\description{
apply GWAlpha method from Fournier-Level et 2017
}
\examples{
sim = PhenoSim (1000, 100, 10, 0.6, 5)
K = 7
clusters = quantile_clustering (data.frame(sim$phenotypes), K)
Freq= compute_group_MAFs(sim$genotypes,as.factor(clusters))
test.stat<-GWalpha(sim$phenotypes[,1], clusters, Freq, K)

}
