% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoSim.R
\name{make_pooled_data_design_matrix}
\alias{make_pooled_data_design_matrix}
\title{make_pooled_data_design_matrix}
\usage{
make_pooled_data_design_matrix(genotypes, phenotypes, k = 2)
}
\arguments{
\item{genotypes}{data frame}

\item{phenotypes}{data frame}

\item{k}{integer}
}
\value{
list
}
\description{
create the design matrix of the pooled data
}
\examples{
sim = PhenoSim (1000, 100, 10, 0.6, 5)
Y<-data.frame(sim$phenotypes)
X<-sim$genotypes
X<-make_pooled_data_design_matrix(X, Y, 7)

}
