% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newgen2phen.R
\name{newgen2phen}
\alias{newgen2phen}
\title{Associate regenerated genotypes to phenotypes}
\usage{
newgen2phen(clusters, regenerate, sim)
}
\arguments{
\item{clusters}{Numeric vector.}

\item{regenerate}{Numeric vector}

\item{sim}{PhenotypeSimulator object}
}
\value{
dataframe of associated genotypes
}
\description{
Associate regenerated genotypes to phenotypes
}
\examples{
sim = sim2snp(100,1000,10)
pools = phen2pool(sim,3,3)
GenotypesCluster = trugen2clust(pools,sim)
freq = pool2freq(pools,GenotypesCluster,sim)
regen = newgenotypes(GenotypesCluster,pools,sim,freq)
newgen_phen_comb = newgen2phen(pools, regen, sim)
}
