% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen2pool.R
\name{phen2pool}
\alias{phen2pool}
\title{Cluster individuals based on their phenotypes into equal-sized-pool for pool-seq}
\usage{
phen2pool(sim, K, L)
}
\arguments{
\item{sim}{a PhenotypeSimulator object.}

\item{K}{Numeric vector}

\item{L}{Numeric vector}
}
\value{
list
}
\description{
The goal is to cluster the individuals based on their phenotypes,
Trait 1 is divided into \code{K} quantiles
Trait2 is divided into \code{L} quantiles
}
\examples{
sim = sim2snp(100,100,10)
pools = phen2pool(sim, 3, 3)
}
