% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoSim.R
\name{PhenoSim}
\alias{PhenoSim}
\title{Simulate phenotypes and genotypes using PhenotypeSimulator}
\usage{
PhenoSim(
  n = 2000,
  SNP = 100,
  n_causal = 10,
  h2 = 0.5,
  p = 5,
  pcorr = 0.8,
  seed = 123
)
}
\arguments{
\item{n}{Numeric vector}

\item{SNP}{Numeric vector.}

\item{n_causal}{Numeric vector}

\item{h2}{Numeric vector}

\item{p}{Numeric vector}

\item{pcorr}{Numeric vector}

\item{seed}{Numeric vector}
}
\value{
a list
}
\description{
PhenotypeSimulator wrapper to generates p-dimensional phenotypes
Total number of individuals \code{n}, default=2000
Total number of SNPs \code{SNP}, default=100
Number of phenotypes \code{p}, default=5
SNP heritability \code{h2s}, default=50%
Number of causal SNPs \code{n_causal}, default=10
Correlation between phenotype \code{pcorr}, default=0.8
Seed for random generaion \code{seed}, default=123
}
\examples{
sim = PhenoSim (1000, 100, 10, 0.6, 5)

}
