% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poogwam.R
\name{poogwam}
\alias{poogwam}
\title{poogwam}
\usage{
poogwam(
  Method = "poogwam",
  Freq,
  phenotypes,
  clusters,
  K = 0,
  eta = 1,
  max_iter = 1000,
  tol = 1e-08,
  seed = 1,
  nb.batch = 1,
  trace = FALSE
)
}
\arguments{
\item{Method}{string}

\item{Freq}{data frame of allele frequencies}

\item{phenotypes}{data frame of phenotypes}

\item{clusters}{pool ID}

\item{K}{interger number of pools}

\item{eta}{default=1}

\item{max_iter}{float default=1e3}

\item{tol}{float default=1e-8}

\item{seed}{default=1}

\item{nb.batch}{float default=1}

\item{trace}{boolean default=FALSE}
}
\value{
numeric vector of SNP effects
}
\description{
GWAS on pool-seq data using several methods
}
\examples{
sim = PhenoSim ( 1000, 100, 10, 0.6, 3 )
K = 4
clusters = quantile_clustering ( data.frame ( sim$phenotypes ), K )
Freq = compute_group_MAFs( sim$genotypes, as.factor(clusters) )
pv = poogwam ("poogwam", Freq, sim$phenotypes[,1], clusters )

}
