% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool2freq.R
\name{pool2freq}
\alias{pool2freq}
\title{Getting Pooled allele frequencies}
\usage{
pool2freq(clusters, clusters_genotype, phenotype)
}
\arguments{
\item{clusters}{a list of clusters}

\item{clusters_genotype}{a list of clusters with genotype.}

\item{phenotype}{a PhenotypeSimulator output}
}
\value{
a list of clusters with allele frequencies
}
\description{
The following function returns the PAF of a paricular SNP
 in all the pools. The output is a vector whose ith element is PAF of
 the snp in ith pool. Argument 'clusters_genotype' should be the
 output of the 'cls_genotype' function. The next argument 'snp' should be
 a character string denoting the SNP, for e.g., 'SNP_111', 'SNP_176' etc.
}
\examples{
sim = sim2snp(100, 1000, 10)
pools = phen2pool(sim, 3, 3)
GenotypesCluster = trugen2clust (pools, sim)
alleleFreq = pool2freq (pools, GenotypesCluster, sim)
}
