% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.from.test.stat.R
\name{pvalues.from.test.stat}
\alias{pvalues.from.test.stat}
\title{pvalues.from.test.stat}
\usage{
pvalues.from.test.stat(test.stat, alpha = 0.01)
}
\arguments{
\item{test.stat}{numeric vector of SNP effects}

\item{alpha}{numeric vector of th error}
}
\value{
numeric vector
}
\description{
get p-values method from Fournier-Level et 2017
}
\examples{
sim = PhenoSim (1000, 100, 10, 0.6, 5)
K = 7
clusters = quantile_clustering (data.frame(sim$phenotypes), K)
Freq= compute_group_MAFs(sim$genotypes,as.factor(clusters))
test.stat<-GWalpha(sim$phenotypes, clusters, Freq, K)
pv = pvalues.from.test.stat (test.stat)

}
