% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_clustering.R
\name{quantile_clustering}
\alias{quantile_clustering}
\title{quantile_clustering}
\usage{
quantile_clustering(data, n_quantiles = 4)
}
\arguments{
\item{data}{data frame of phenotypes}

\item{n_quantiles}{numeric vector for the number of quantiles for each phenotype}
}
\value{
list
}
\description{
Cluster individuals based on their phenotypes into equal-sized pool for pool-seq
}
\examples{
par(mfrow=c(1,2))
clusters <- quantile_clustering(y<-data.frame(matrix(rnorm(500),500,1)))
plot(data.frame(1:500,y),col=clusters,main="1D phenotypes",xlab="Index",ylab="Trait 1")
clusters <- quantile_clustering(y<-data.frame(matrix(rnorm(1000),500,2)))
plot(y,col=clusters,main="2D phenotypes",xlab="Trait 1",ylab="Trait 2")

}
