% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection}
\alias{selection}
\title{Detection of causal SNPs}
\usage{
selection(data)
}
\arguments{
\item{data}{phenotypes and genotypes data}
}
\value{
a dataframe of causal SNP
}
\description{
The goall is to discover causal SNP from regenrated genotypes data and multitrait data
}
\examples{
sim = sim2snp(100,1000,10)
pools = phen2pool(sim,3,3)
GenotypesCluster = trugen2clust(pools,sim)
freq = pool2freq(pools,GenotypesCluster,sim)
regen = newgenotypes(GenotypesCluster,pools,sim,freq)
newgen_phen_comb = newgen2phen(pools, regen, sim)
causal_snps = selection (newgen_phen_comb)
}
