% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2snp.R
\name{sim2snp}
\alias{sim2snp}
\title{Simulate phenotypes and genotypes}
\usage{
sim2snp(n, tNrSNP, cNrSNP, p = 2, gvar = 0.6, snph2 = 0.5, pleio = 0.2)
}
\arguments{
\item{n}{Numeric vector.}

\item{tNrSNP}{Numeric vector}

\item{cNrSNP}{Numeric vector}

\item{p}{Numeric vector}

\item{gvar}{Numeric vector}

\item{snph2}{Numeric vector}

\item{pleio}{numeric vector}
}
\value{
a PhenotypeSimulator object
}
\description{
PhenotypeSimulator wrapper to generates p-dimensional phenotypes
with 3 components: genetic effects, polygenic effects, and environmental noise.
There is no covariaye effect (hence 'delta'=0) and no correlated noise (hence 'rho'=0).
Total number of individuals \code{n}, default=2000,
Number of phenotypes \code{p}, default = 2
total number of SNPs \code{tNrSNP}, default=10000,
Number of causal SNPs \code{cNrSNP}, default=10.
Total genetic heritability \code{gvar}, default=60%,
SNP heritability * \code{h2s}, default=50%
Ratio of Pleitropic SNPs \code{pleio}, default=0.2
We randomly select 2 SNPs out of 10  to be causal for both the traits
(i.e. pleitropic SNPs or SNPs having shared effect);
and this quantity is controlled by '1-pIndependerGenetic'.
So we set '1-pIndependerGenetic'=2/10=0.2 or 'pIndependerGenetic'=0.8.
'pTraitIndependentGenetic = 0.5' means that the remaining 8 SNPs are randomly selected to be causal for only one trait (i.e 1 out of 2 traits. Hence 1/2=0.5 appears)
'SNPfrequencies' denote the set of MAFs. MAF of a particular SNP is randomly selected from this set of MAFs.
}
\examples{
sim = sim2snp(100, 100, 10)
}
