% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singValRidge.R
\name{singValRidge}
\alias{singValRidge}
\title{singValRidge}
\usage{
singValRidge(phenotypes, Freq, clusters, K)
}
\arguments{
\item{phenotypes}{data frame of phenotypes}

\item{Freq}{numeric vector of allele frequencies}

\item{clusters}{vector of cluster ID}

\item{K}{number of clusters}
}
\value{
numeric vector
}
\description{
apply singular value decomposition for ridge analysis
}
\examples{
sim <- PhenoSim (1000, 100, 10, 0.6, 5)
clusters <- quantile_clustering (data.frame(sim$phenotypes))
Freq <- compute_group_MAFs(sim$genotypes, as.factor(clusters))
criterion <- singValRidge(sim$phenotypes, Freq, clusters, K = 4)

}
