% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenGWASwrap.R
\name{statgenGWASwrap}
\alias{statgenGWASwrap}
\title{statgenGWASwrap}
\usage{
statgenGWASwrap(genotypes, phenotypes)
}
\arguments{
\item{genotypes}{data frame of genotypes}

\item{phenotypes}{data frame of phenotypes}
}
\value{
statgenGWAS p-values
}
\description{
a wrapper for the statgenGWAS package to run on simulated data
}
\examples{
sim = PhenoSim ( 100, 100, 10, 0.6, 2 )
pv = statgenGWASwrap ( sim$genotypes, sim$phenotypes )

}
