% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trugen2clust.R
\name{trugen2clust}
\alias{trugen2clust}
\title{Assigning true genotypes to pools}
\usage{
trugen2clust(clusters, sim)
}
\arguments{
\item{clusters}{cluster of individual}

\item{sim}{a PhenotypeSimulator object}
}
\value{
genotypes assigned to cluster
}
\description{
This will assign the true genotypes to the cluster based on phenotypes
}
\examples{
sim = sim2snp(100,1000,10)
pools = phen2pool(sim, 3, 3)
GenotypesCluster = trugen2clust (pools, sim)
}
