% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chisquare.R
\name{Chisquare}
\alias{Chisquare}
\title{Chisquare}
\usage{
Chisquare(Freq, clusters)
}
\arguments{
\item{Freq}{data frame of genotypes}

\item{clusters}{integers}
}
\value{
list
}
\description{
apply classical chi-square
}
\examples{
sim = PhenoSim (500, 100, 10, 0.7, 2)
clusters = quantile_clustering (data.frame(sim$phenotypes), 4)
Freq = compute_group_MAFs( sim$genotypes, as.factor(clusters) )
pv <- Chisquare (Freq, clusters)


}
