% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAlpha.R
\name{GWalpha}
\alias{GWalpha}
\title{GWAlpha}
\usage{
GWalpha(
  phenotypes = phenotypes,
  d = 1,
  clusters = clusters,
  Freq = Freq,
  penalized = TRUE
)
}
\arguments{
\item{phenotypes}{data frame of phenotypes}

\item{d}{phenotypes index to be analyzed}

\item{clusters}{numeric vector of the clusters}

\item{Freq}{numeric vector of allele frequencies}

\item{penalized}{boolean}
}
\value{
numeric vector
}
\description{
apply the GWAlpha algorithm from Fournier-Level et al. 2017
}
\examples{
sim = PhenoSim (1000, 100, 10, 0.6, 5)
clusters = quantile_clustering (data.frame(sim$phenotypes))
Freq= compute_group_MAFs(sim$genotypes,as.factor(clusters))
test.stat<-GWalpha(sim$phenotypes,d=1, clusters, Freq)

}
